Start implementation of BlockBasedIndexPopulator
BlockBasedIndexPopulator extends NativeIndexPopulator to reuse functionality
typical for native populator such as creating GBPTree and cleaning up files.

Important things that are missing:
- Drop
- VerifyDeferredConstraints
- Sampling
- Close
- newReader
- Handle external updates after building the tree from block storage.

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/index/schema/ByteBufferFactory.java
@@ -23,5 +23,5 @@ import java.nio.ByteBuffer;
 
 interface ByteBufferFactory
 {
-    ByteBuffer newBuffer( long bufferSize );
+    ByteBuffer newBuffer( int bufferSize );
 }
